skips downloading interview mfcc1 if already there

+++ tasks/SEAME/preprocess/get_mfcc.sh
@@ -1,14 +1,20 @@
 #!/bin/bash
 
 cd ..
-mkdir -p data/interview/mfcc
-cd data/interview/mfcc
-wget http://tts.speech.cs.cmu.edu/rsk/misc_stuff/feats_seame_interview_07march2019.tar.gz
-tar -xvzf feats_seame_interview_07march2019.tar.gz
-mv feats_interview/cleaned ..
-cd ..
-rm -rf mfcc
-mv cleaned mfcc1
+
+if [ -d "data/interview/mfcc1" ] && [[ $(ls -A data/interview/mfcc1 | wc -l) -eq 40600 ]]; then
+    cd data/interview
+else
+    mkdir -p data/interview/mfcc
+    cd data/interview/mfcc
+    wget http://tts.speech.cs.cmu.edu/rsk/misc_stuff/feats_seame_interview_07march2019.tar.gz
+    tar -xvzf feats_seame_interview_07march2019.tar.gz
+    mv feats_interview/cleaned ..
+    cd ..
+    rm -rf mfcc
+    mv cleaned mfcc1
+fi
+
 cd ..
 mkdir -p conversation/mfcc
 cd conversation/mfcc
